Added comments
Also improved how shading is calculated slightly. Now see if we are in shadow and don't calculate shade of pixel if we are, before calculating shade for every light and then overriding it to be black if it was in shade. Should result in less shading calculations. At low fps (around 10) improved performance by 1fps roughly, didn't get a before reading at higher fps.